add base project structure
