+++ lib/ui/app.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class App extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Tortiki',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+      ),
+      home: Container(),
+    );
+  }
+}
